Rename seeds command to commits
