+++ client/commands/commits/export.go
@@ -1,4 +1,4 @@
-package seeds
+package commits
 
 import (
 	"github.com/pkg/errors"
@@ -12,21 +12,21 @@ import (
 
 var exportCmd = &cobra.Command{
 	Use:   "export <file>",
-	Short: "Export selected seeds to given file",
-	Long: `Exports the most recent seed to a binary file.
+	Short: "Export selected commits to given file",
+	Long: `Exports the most recent commit to a binary file.
 If desired, you can select by an older height or validator hash.
 `,
-	RunE:         commands.RequireInit(exportSeed),
+	RunE:         commands.RequireInit(exportCommit),
 	SilenceUsage: true,
 }
 
 func init() {
-	exportCmd.Flags().Int(heightFlag, 0, "Show the seed with closest height to this")
-	exportCmd.Flags().String(hashFlag, "", "Show the seed matching the validator hash")
+	exportCmd.Flags().Int(heightFlag, 0, "Show the commit with closest height to this")
+	exportCmd.Flags().String(hashFlag, "", "Show the commit matching the validator hash")
 	RootCmd.AddCommand(exportCmd)
 }
 
-func exportSeed(cmd *cobra.Command, args []string) error {
+func exportCommit(cmd *cobra.Command, args []string) error {
 	if len(args) != 1 || len(args[0]) == 0 {
 		return errors.New("You must provide a filepath to output")
 	}
